feat: simplified bun dockerfile
