+++ _bun/types/go.d.ts
@@ -1,0 +1,13 @@
+/**
+ * Go is the class as defined in the Golang `wasm_exec.js` distributable file required for WebAssembly.
+ * Golang WebAssembly wiki: https://github.com/golang/go/wiki/WebAssembly
+ */
+declare class Go {
+	argv: string[];
+	env: { [envKey: string]: string };
+	exit: (code: number) => void;
+	importObject: WebAssembly.Imports;
+	exited: boolean;
+	mem: DataView;
+	run(instance: WebAssembly.Instance): Promise<void>;
+}
